module 1 fix

+++ module1/src/main/java/com/alevel/java/nix/level1/task3/AreaOfTrian.java
@@ -2,7 +2,7 @@ package com.alevel.java.nix.level1.task3;
 
 public class AreaOfTrian {
     public double areaOfTriangle(Point pointA, Point pointB, Point pointC){
-        float area;
+        double area;
 
         area = Math.abs(pointA.getX() * (pointB.getY() - pointC.getY())
                 + pointB.getX() * (pointC.getY() - pointA.getY())
